moved ifOnFishCollect to utils
