+++ client/script/utils/ifOnFishCollect.js
@@ -1,10 +1,10 @@
 /* eslint-disable guard-for-in */
 import {hitTestRectangle} from './PIXIutils'
-import store, {updateFish, removeFish} from '../store'
-import {stage} from './game'
+import store, {updateFish, removeFish} from '../../store'
+import {stage} from '../game'
 import {Text} from 'pixi.js'
-import socket from '../socket'
-import {TILE_SIZE} from './CONSTANTS'
+import socket from '../../socket'
+import {TILE_SIZE} from '../CONSTANTS'
 
 export const ifOnFishCollect = (boat, fishes) => {
   // check how many fishes the boat has/can take in
